+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## 1.3.2 - Ronin's Bubble
-- Replaced moving description in items by a big popup on hover.
+- Replaced moving description in items by a big tooltip on hover.
 - Added PoW Revised 20Q (for old Ronin, just modify the template in 20Q to change the character sheet fields).
 - Added Bushido tenets on Social tab in sheet for all (I think this is useful for players, and it's in Core rulebook so... why this is not by default ;) ).
 - Added English/French Journal Compendiums for Conditions and Terrain Qualities (thanks to TesserWract).
@@ -9,6 +9,7 @@
 - Added PoW Name Tables compendium : Qamarist Names, Ivory Kingdoms Names, Rokugani Names and Ujik Names.
 - Added ability to drop a (Curriculum) Journal to have a button in school experience tab.
 - Added properties antagonists (ex : Sacred remove Unholy on drop).
+- Added when switching between NPC types, now the prototype token is altered this way : Linked actor for Adversary, unlinked for Minion.
 - Fixed school compendium: "Kuni Warden School" from "Scorpion" to "Crab".
 - Fixed roll for RollTable not showing the text.
 - Fixed some Pdf Copy-Paste error in school cursus that make L instead of I: Lkoma/Lsawa/Luchi...